Change type of player_t::action_list_id_ from size_t to unsigned and remove the unnecessary cast. Don't complain about the number of action lists until creating the 65th one instead of after creating the 64th.

+++ engine/simulationcraft.hpp
@@ -5255,7 +5255,7 @@ public:
 
   // Internal counter for action priority lists, used to set
   // action_priority_list_t::internal_id for lists.
-  size_t action_list_id_;
+  unsigned action_list_id_;
 
   // Figure out another actor, by name. Prioritizes pets > harmful targets >
   // other players. Used by "actor.<name>" expression currently.
